refactor: reorganize trip package structure
