+++ src/main/java/com/chocobi/leafy/distance/service/TransDistanceService.java
@@ -7,8 +7,8 @@ import com.chocobi.leafy.distance.domain.TransDistanceRequest;
 import com.chocobi.leafy.distance.domain.TransDistanceBatchRequest;
 import com.chocobi.leafy.distance.dto.*;
 import com.chocobi.leafy.place.application.PlaceService;
-import com.chocobi.leafy.trip.service.TripPlaceService;
-import com.chocobi.leafy.trip.dto.TripPlaceResponse;
+import com.chocobi.leafy.trip.application.TripPlaceService;
+import com.chocobi.leafy.trip.dto.response.TripPlaceResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
