+++ src/main/java/com/chocobi/leafy/trip/dto/request/TripPlaceRequest.java
@@ -1,4 +1,4 @@
-package com.chocobi.leafy.trip.dto;
+package com.chocobi.leafy.trip.dto.request;
 
 import lombok.Data;
