+++ src/main/java/com/chocobi/leafy/trip/infra/TripCommandService.java
@@ -1,0 +1,4 @@
+package com.chocobi.leafy.trip.infra;
+
+public class TripCommandService {
+}
