+++ src/main/java/com/chocobi/leafy/trip/infra/TripFindService.java
@@ -1,0 +1,4 @@
+package com.chocobi.leafy.trip.infra;
+
+public class TripFindService {
+}
