+++ src/main/java/com/chocobi/leafy/trip/infra/entity/Trip.java
@@ -1,4 +1,4 @@
-package com.chocobi.leafy.trip.entity;
+package com.chocobi.leafy.trip.infra.entity;
 
 import com.chocobi.leafy.place.infra.entity.RegionGroup;
 import com.chocobi.leafy.user.infra.entity.UserEntity;
