+++ src/main/java/com/chocobi/leafy/trip/presentation/TripDocs.java
@@ -1,0 +1,4 @@
+package com.chocobi.leafy.trip.presentation;
+
+public interface TripDocs {
+}
