+++ src/main/java/com/chocobi/leafy/trip/vo/TripError.java
@@ -1,0 +1,4 @@
+package com.chocobi.leafy.trip.vo;
+
+public enum TripError {
+}
